packages/puck/puck-1.0.7-py2.py3-none-any.whl/puck/test_parser.py
from puck.backend import echo_backend
from puck.parser import parse_requirement


def test_parse_req():
    expected = {
        'name': 'foo',
        'pinned_version': '1.2.3',
        'latest_version': 'foo',
        'source': None
    }
    assert parse_requirement(
        'foo==1.2.3',
        None,
        backend=echo_backend
    ) == expected


def test_parse_with_spaces():
    expected = {
        'name': 'foo',
        'pinned_version': '1.2.3',
        'latest_version': 'foo',
        'source': None
    }
    assert parse_requirement(
        'foo == 1.2.3',
        None,
        backend=echo_backend
    ) == expected


def test_parse_invalid_req():
    assert parse_requirement(
        'foo<1.2.3',
        None,
        backend=echo_backend
    ) is None
